hyprpm: cast std::clamp args to be of the same type
hyprpm/src/progress/CProgressBar.cpp:50:27: error: no matching function for call to 'clamp'
    const auto BARWIDTH = std::clamp(w.ws_col - m_szCurrentMessage.length() - 2, 0UL, 50UL);
                          ^~~~~~~~~~
/usr/include/c++/v1/__algorithm/clamp.h:38:1: note: candidate template ignored: deduced conflicting types for parameter '_Tp' ('size_type' (aka 'unsigned int') vs. 'unsigned long')
clamp(const _Tp& __v, const _Tp& __lo, const _Tp& __hi)
^
/usr/include/c++/v1/__algorithm/clamp.h:27:1: note: candidate function template not viable: requires 4 arguments, but 3 were provided
clamp(const _Tp& __v, const _Tp& __lo, const _Tp& __hi, _Compare __comp)
^

+++ hyprpm/src/progress/CProgressBar.cpp
@@ -47,7 +47,7 @@ void CProgressBar::print() {
     else
         percentDone = (float)m_iSteps / (float)m_iMaxSteps;
 
-    const auto BARWIDTH = std::clamp(w.ws_col - m_szCurrentMessage.length() - 2, 0UL, 50UL);
+    const auto BARWIDTH = std::clamp(w.ws_col - static_cast<unsigned long>(m_szCurrentMessage.length()) - 2, 0UL, 50UL);
 
     // draw bar
     message += std::string{" "} + Colors::GREEN;
